Increase size of working arrays in the makekeys utility program.
Makekeys is used to create an optimal hash of the keysyms defined
in x11proto’s keysymdef.h.

The recent addition of new keysyms there has triggered a bug in
makekeys where it tries to use a zero on the rhs of the % (mod)
operator (resulting in a divide by zero error) whenever it fails
to find a solution within its constraints.

Increasing the size of the arrays allows it to find a solution for
the current set of keysyms.

Makekeys is only run durring the build process, so this has no impact
on users of libX11, only on the amount of VM needed to build it.

It still needs a more complete fix, but this allows compiles to
progress until that is completed.