XIM: Allow modifier releases in compose sequences (#26705)
Currently, only non-modifier keys (actually, keysyms) can be part of a compose
sequence, and they are matched against the defined compose sequences at the
time the key is pressed.  The patch allows to use modifier keys an well, but
matches them on key release, and only if no other key has been pressed after
the modifier.

Releasing a non-matched modifier during an ongoing compose sequence only aborts
the sequence if any modifier release would have matched.  In particular, if no
compose sequences with modifiers are specified, the compose mechanism works
exactly as without this patch.

Even if modifiers are part of a compose sequence, they are not filtered.  This
is because modifiers affect the keyboard state no matter what we do here and,
therefore, filtering them only could confuse clients.

The purpose is this extension to the compose mechanism is to allow to make
better use of keys in convenient reach for touch typing.

Signed-off-by: Andreas Wettstein <wettstein509@solnet.ch>
Signed-off-by: James Cloos <cloos@jhcloos.com>