Align files with upstream libX11-1.6.0
This commit looks a bit crazy at first glance. It (re-)introduces lots
of whitespaces and bad formatting.

Explanation: Backporting upstream changes lead to commits being
applied out of order. This meant a lot of manual intervention which in
turn lead to slight differences between upstream and NX.

With this commit these slight differences are minimized which will be
of great help when adding further upstream patches.