From 7f8f9a36ef901f31279c385caf960a22daeb33fe Mon Sep 17 00:00:00 2001 From: "Owen W. Taylor" <otaylor@fishsoup.net> Date: Fri, 9 May 2014 18:21:05 -0400 Subject: [PATCH] Fix XNextRequest() after direct usage of XCB
When XCB owns the X socket, dpy->request is not updated, so
NextRequest() and XNextRequest() return the wrong value. There's
nothing we can do to fix NextRequest() while retaining ABI compat,
but change XNextRequest() to grab the socket back from XCB,
updating dpy->request.

Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>
Reviewed-by: Uli Schlachter <psychon@znc.in>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>

Backported-to-NX-by: Ulrich Sibiller <uli42@gmx.de>:

Added #ifdefs to be aware of changes regarding XCB in case we later
switch to XCB.