Free xkb devPrivates on CloseDevice
The situation here is unclear. There are chances that this needs to be
revised later.

With the integration of core keyboard and pointer memory leaks started
to (re-)appear.

For the core devices XkbSetExtension() is called on initialization and
allocates private data. But there's no corresponding code to free the
private data on CloseDevice. So I added XkbFreePrivates(). This
function should be called on DEVICE_CLOSE.

For the NX devices XkbSetExtension() is never called so the privates
are not allocated which _feels_ wrong but I am not sure if it really
_is_ wrong.

Nevertheless let's call XkbFreePrivates() for both core and NX devices
including sanity checks and nullification to be on the safe side.