Atoms.c: fix nesting detection
Fixes ArcticaProject/nx-libs#663

The previous logic was completely broken as it depended on the order
of atoms being created on the real X server. This might have been
working in the beginning of NX but eventually atom mapping between
nxagent and the real X server was introduced which broke the whole
logic. As a rather ugly fix let's not create the NX_IDENTITY and
NX_AGENT_SIGNATURE atoms to the real X server. If we find them
nevertheless on the remote side we are running nested.

This change is neither nice nor efficient, just functional. It should
be redesigned one day.