Clipboard.c: fix: remove wrong optimization
In the "none" case we must ensure the clipboard data is reset. This
"optimization" skipped the reset part which made the code think it is
waiting for an answer. Which (of course) never arrived. So any further
request was rejected...

Also reduce debugging output for the "none" case because
resetClientSelectionStage will print it anyway.