Clipboard.c: implement target caching
Some applications (e.g. Chrome, Firefox) send multiple TARGETS
requests when they want to paste a selection (reason currently
unknown). If such an application runs in an NX session and the current
clipboard owner is on the real X server this triggers many
(superflous) roundtrips.

By caching the first answer to those requests and answering subsequent
requests from the cache we can eliminate all this communication.

The implemention resulted from attempting to fix
ArcticaProject/nx-libs/issues/#969 (solution 1a) but it turned out
that the speedup this brings does not help in reducing the number of
requests by firefox. It must be a bug in the firefox code.

Nevertheless I kept the code as it should speed up the whole pasting
process considerably while reducing communication over the wire.

This also eliminates a memory leak in the target forwarding code which
was unnoticed before.