Screen.c: remove superfluous fixme
As we do not have PictureScreenPrivateIndex anymore we cannot (un)set it.

I cannot think of a proper alternative. Either the replacement
PictureScreenPrivateKey exists or not, in that latter case it
will be created automatically on dixLookupPrivate().

The index used to be checked in the GetPictureScreenIfSet() macro.
this is now just mapped to GetPictureScreen() which does the
dixLookupPrivate().