NXglyph.c: make nxagentAddGlyphs work again
Due to changes in Xorg a glyph is no longer followed by the bits to
draw but by several PicturePtrs (one per Screen). We must have a
ZPixmap to pass it on to nxagentAddGlyphs() resp. XRenderAddGlyphs().

I am not sure if that is the optimal way to achieve the goal but for
now it is sufficient!

Also the size calculation is working but might be problematic for
other combinations of depth and size.